+++ ChangeLog.md
@@ -2,10 +2,11 @@
 
 ## 0.27.1
 ### 2024-04-15
-
 * FIX: multi-asset run was failure due to including schedule cashflow run.
 * ENHANCE: Enable revovling on `Receivable`
 * ENHANCE: add `RecoveryByDays` to `Receivable` ,which describle recovery cash received after default.
+* BUG: fix single asset run on `lease`
+
 
 ## 0.27.0
 ### 2024-04-01